Refactor watershed panel to handle sub watershed panel components.
