+++ client/src/routes/router.tsx
@@ -1,7 +1,9 @@
 import { lazyRouteComponent, createRoute, createRootRoute, createRouter, Outlet } from '@tanstack/react-router';
 import Navbar from '../components/navbar/Navbar';
 import Home from '../pages/home/Home';
-import Watershed from '../components/watershed/Watershed';
+import Watershed from '../components/side-panels/watershed/WatershedPanel';
+import WatershedOverview from '../components/side-panels/watershed/WatershedOverview';
+import WatershedDataPanel from '../components/side-panels/watershed/WatershedDataPanel';
 
 const About = lazyRouteComponent(() => import('../pages/about/About'));
 const FAQ = lazyRouteComponent(() => import('../pages/faq/FAQ'));
@@ -19,7 +21,7 @@ const rootRoute = createRootRoute({
       </div>
     </>
   ),
-  notFoundComponent: () => <div>Page not found</div>,
+  notFoundComponent: () => <div>404: Page Not Found</div>,
 });
 
 const homeRoute = createRoute({
@@ -32,6 +34,18 @@ const watershedIdRoute = createRoute({
   getParentRoute: () => homeRoute,
   path: '/watershed/$webcloudRunId',
   component: Watershed,
+});
+
+const watershedOverviewRoute = createRoute({
+  getParentRoute: () => watershedIdRoute,
+  path: '/',
+  component: WatershedOverview,
+});
+
+const watershedDataRoute = createRoute({
+  getParentRoute: () => watershedIdRoute,
+  path: 'data',
+  component: WatershedDataPanel,
 });
 
 const aboutRoute = createRoute({
@@ -60,7 +74,9 @@ const loginRoute = createRoute({
 
 // Create the route tree
 const routeTree = rootRoute.addChildren([
-  homeRoute.addChildren([watershedIdRoute]),
+  homeRoute.addChildren([
+    watershedIdRoute.addChildren([watershedOverviewRoute, watershedDataRoute]),
+  ]),
   aboutRoute,
   faqRoute,
   documentationRoute,
